update field documentation, change file for fieldFreeFunctions, reorder basics a bit
