+++ doc/basics/fields.rst
@@ -22,9 +22,11 @@ Besides these container like field classes several finite volume specific field 
 
 The Field<ValueType> class
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
-The Field classes are the central elements for implementing a platform portable CFD framework. Fields should allow to perform basic algebraic operations such as binary operations like the addition or subtraction of two fields, or scalar operations like the multiplication of a field with a scalar.
+The Field class is the basic container class and is the central elements for implementing a platform portable CFD framework.
+Fields should allow to perform basic algebraic operations such as binary operations like the addition or subtraction of two fields, or scalar operations like the multiplication of a field with a scalar.
 
-In the following, we will explain the implementation details of the field operations using the additions operator as an example. The block of code below shows an example implementation of the addition operator.
+In the following, some implementation details of the field operations are detailed using the additions operator as an example.
+The block of code below shows an example implementation of the addition operator.
 
 .. code-block:: cpp
 
@@ -37,34 +39,49 @@ In the following, we will explain the implementation details of the field operat
     }
 
 
-Besides creating a temporary for the result it mainly calls the free standing ``add`` function which is implemented in ``FieldOperations.hpp``. This, in turn, dispatches to the ``addOp`` functor, that holds the actual addition kernels. In the case of addition this is implemented as a  ``Kokkos::parallel_for`` function, see `Kokkos documentation  <https://kokkos.org/kokkos-core-wiki/API/core/parallel-dispatch/parallel_for.html>`_ for more details.
+Besides creating a temporary for the result it mainly calls the free standing ``add`` function which is implemented in ``include/NeoFOAM/field/fieldFreeFunctions.hpp``.
+This, in turn, dispatches to the ``fieldBinaryOp`` function, passing the actual kernel as lambda.
+The ``fieldBinaryOp``  is implemented using our parallelFor implementations which ultimately dispatch to the ``Kokkos::parallel_for`` function, see `Kokkos documentation  <https://kokkos.org/kokkos-core-wiki/API/core/parallel-dispatch/parallel_for.html>`_ for more details.
 
 .. code-block:: cpp
 
-   using executor = typename Executor::exec;
-   auto a_f = a.field();
-   auto b_f = b.field();
-   Kokkos::parallel_for(
-      Kokkos::RangePolicy<executor>(0, a_f.size()),
-      KOKKOS_CLASS_LAMBDA(const int i) { a_f[i] = a_f[i] + b_f[i]; }
-   );
+    template<typename ValueType>
+    void add(Field<ValueType>& a, const Field<std::type_identity_t<ValueType>>& b)
+    {
+      detail::fieldBinaryOp(
+          a, b, KOKKOS_LAMBDA(ValueType va, ValueType vb) { return va + vb; }
+      );
+    }
 
-The code snippet also highlights another important aspect, the executor. The executor, here defines the ``Kokkos::RangePolicy``, see  `Kokkos Programming Model  <https://github.com/kokkos/kokkos-core-wiki/blob/main/docs/source/ProgrammingGuide/ProgrammingModel.md>`_. Besides defining the RangePolicy, the executor also holds functions for allocating and deallocationg memory. A full example of using NeoFOAMs fields with a GPU executor could be implemented as
-
+A simplified version of the ``parallelFor`` function is shown below.
 .. code-block:: cpp
+    template<typename Executor, parallelForKernel Kernel>
+    void parallelFor(
+        [[maybe_unused]] const Executor& exec, std::pair<size_t, size_t> range, Kernel kernel
+    )
+    {
+        auto [start, end] = range;
+        if constexpr (std::is_same<std::remove_reference_t<Executor>, SerialExecutor>::value)
+        {
+        ...
+        }
+        else
+        {
+            using runOn = typename Executor::exec;
+            Kokkos::parallel_for(
+                "parallelFor",
+                Kokkos::RangePolicy<runOn>(start, end),
+                KOKKOS_LAMBDA(const size_t i) { kernel(i); }
+            );
+        }
+    }
 
-    NeoFOAM::GPUExecutor GPUExec {};
-    NeoFOAM::Field<NeoFOAM::scalar> GPUa(GPUExec, N);
-    NeoFOAM::fill(GPUa, 1.0);
-    NeoFOAM::Field<NeoFOAM::scalar> GPUb(GPUExec, N);
-    NeoFOAM::fill(GPUb, 2.0);
-    auto GPUc = GPUa + GPUb;
+The code snippet highlights another important aspect, the executor.
+The executor defines the ``Kokkos::RangePolicy``, see  `Kokkos Programming Model  <https://github.com/kokkos/kokkos-core-wiki/blob/main/docs/source/ProgrammingGuide/ProgrammingModel.md>`_.
+Besides defining the RangePolicy, the executor also holds functions for allocating and deallocationg memory.
+See our `documentation  <https://exasim-project.com/NeoFOAM/latest/basics/executor.html>`_ for more details on the executor model.
 
-Interface
-^^^^^^^^^
-
-.. doxygenclass:: NeoFOAM::Field
-    :members:
+Further `Details  <https://exasim-project.com/NeoFOAM/latest/doxygen/html/classNeoFOAM_1_1Field.html>`_.
 
 Cell Centered Fields
 ^^^^^^^^^^^^^^^^^^^^
