+++ include/NeoFOAM/fields.hpp
@@ -6,7 +6,3 @@
 #include "fields/domainField.hpp"
 #include "fields/field.hpp"
 #include "fields/fieldTypeDefs.hpp"
-
-#include "fields/operations/comparison.hpp"
-#include "fields/operations/operationsMacros.hpp"
-#include "fields/operations/sum.hpp"
