+++ include/NeoFOAM/fields/fieldFreeFunctions.hpp
@@ -3,6 +3,7 @@
 #pragma once
 
 #include <tuple>
+#include <span>
 
 #include <Kokkos_Core.hpp>
 #include "NeoFOAM/core/primitives/label.hpp"
@@ -100,6 +101,64 @@ template<typename... Args>
 auto copyToHosts(Args&... fields)
 {
     return std::make_tuple(fields.copyToHost()...);
+}
+
+template<typename T>
+bool equal(Field<T>& field, T value)
+{
+    auto hostField = field.copyToHost();
+    auto hostSpan = hostField.span();
+    for (size_t i = 0; i < hostSpan.size(); i++)
+    {
+        if (hostSpan[i] != value)
+        {
+            return false;
+        }
+    }
+    return true;
+};
+
+template<typename T>
+bool equal(const Field<T>& field, const Field<T>& field2)
+{
+    auto [hostField, hostField2] = copyToHosts(field, field2);
+    auto [hostSpan, hostSpan2] = spans(hostField, hostField2);
+
+    if (hostSpan.size() != hostSpan2.size())
+    {
+        return false;
+    }
+
+    for (size_t i = 0; i < hostSpan.size(); i++)
+    {
+        if (hostSpan[i] != hostSpan2[i])
+        {
+            return false;
+        }
+    }
+
+    return true;
+};
+
+template<typename T>
+bool equal(const Field<T>& field, std::span<T> span2)
+{
+    auto hostSpan = field.copyToHost().span();
+
+    if (hostSpan.size() != span2.size())
+    {
+        return false;
+    }
+
+    for (size_t i = 0; i < hostSpan.size(); i++)
+    {
+        if (hostSpan[i] != span2[i])
+        {
+            return false;
+        }
+    }
+
+    return true;
 }
 
 } // namespace NeoFOAM
